pyproject.toml
[tool.poetry]
name = "earthquake-analysis"
version = "0.1.0"
description = "Analysis of earthquake experimental setup"
authors = ["Martin Esparza <martinesparza22.me@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
jupyterlab = "^4.3.3"
tqdm = "^4.67.1"
matplotlib = "^3.10.0"
seaborn = "^0.13.2"
xarray = "^2024.11.0"
spikeinterface = "^0.101.2"
scikit-learn = "^1.6.1"
pyaldata = { git = "git@github.com:BeNeuroLab/PyalData.git", develop = true }
statsmodels = "^0.14.4"
cupy-cuda12x = "^13.4.1"
torch = "^2.6.0"
opencv-python = "^4.11.0.86"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
